Simple keys are working again
There is a new mechanism to dynamically add/remove the regex/action pair
needed to release a key. This is needed to avoid conflicts across layers.
E.g. if 'A' is assigned to switch.00 on layer0 and 'B' is assigned to the
same switch on layer1, 'B' could claim '!switch.00', preventing the
release of 'A'.

Signed-off-by: Roming22 <roming22@gmail.com>