Scaffolding for parsing/loading combos
A possible improvement around the current implementation could be
to support a complex declaration such as:
  sw.1+( (sw.2|sw.3) & sw.4)
where '+' denotes a non-ordered sequence, '&' denotes an ordered
sequence and '|' denotes an option between multiple switches.
In the above example sw.1 followed by sw.3 followed by sw.4 is a
valid sequence, and so si sw.2, sw.4, sw.1.
However sw.1, sw.4, sw.2 is not.

Signed-off-by: Romain Arnaud <rarnaud@redhat.com>